+++ test/basepack_test_app/app/models/developer.rb
@@ -1,0 +1,6 @@
+class Developer < ActiveRecord::Base
+  # stub name
+  def name
+    "Bill Gates"
+  end
+end