+++ test/basepack_test_app/app/models/issue.rb
@@ -1,0 +1,6 @@
+class Issue < ActiveRecord::Base
+  attr_accessible :developer_id, :project_id, :status_id, :title
+  belongs_to :developer
+  belongs_to :project
+  belongs_to :status
+end