+++ test/basepack_test_app/app/models/status.rb
@@ -1,0 +1,3 @@
+class Status < ActiveRecord::Base
+  # attr_accessible :title, :body
+end