+++ test/basepack_test_app/db/migrate/20130105145110_create_issues.rb
@@ -1,0 +1,12 @@
+class CreateIssues < ActiveRecord::Migration
+  def change
+    create_table :issues do |t|
+      t.string :title
+      t.integer :developer_id
+      t.integer :project_id
+      t.integer :status_id
+
+      t.timestamps
+    end
+  end
+end