+++ test/basepack_test_app/db/migrate/20130105145156_create_projects.rb
@@ -1,0 +1,9 @@
+class CreateProjects < ActiveRecord::Migration
+  def change
+    create_table :projects do |t|
+      t.string :title
+
+      t.timestamps
+    end
+  end
+end