+++ test/basepack_test_app/db/migrate/20130105145231_create_developers.rb
@@ -1,0 +1,8 @@
+class CreateDevelopers < ActiveRecord::Migration
+  def change
+    create_table :developers do |t|
+
+      t.timestamps
+    end
+  end
+end