+++ test/basepack_test_app/app/components/grid_with_in_form_editing_only.rb
@@ -1,0 +1,7 @@
+class GridWithInFormEditingOnly < Netzke::Basepack::Grid
+  def configure(c)
+    super
+    c.model = 'Book'
+    c.enable_edit_inline = false
+  end
+end