+++ test/basepack_test_app/app/components/grid_with_disabled_editing.rb
@@ -1,0 +1,8 @@
+class GridWithDisabledEditing < Netzke::Basepack::Grid
+  def configure(c)
+    super
+    c.model = 'Book'
+    c.enable_edit_inline = false
+    c.enable_edit_in_form = false
+  end
+end