+++ test/basepack_test_app/app/components/grid_with_inline_editing_only.rb
@@ -1,0 +1,7 @@
+class GridWithInlineEditingOnly < Netzke::Basepack::Grid
+  def configure(c)
+    super
+    c.model = 'Book'
+    c.enable_edit_in_form = false
+  end
+end