+++ CHANGELOG.md
@@ -3,6 +3,9 @@
   * Add grid column config option `filter_with` (by firemind)
   * Add ActionColumn (move over from Communitypack)
   * A `column` declaration now appends the declared column to the list of default columns
+  * Add support for autoSync store option (by scho)
+  * Fix a glitch with Grid context menu
+  * Add `enable_edit_inline` option for Grid; when set to false, all editing is done via forms
 
 * bug fix
   * Fix one-to-many association method falling back to 'id'