+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * Fix a glitch with Grid context menu
   * Add `enable_edit_inline` option for Grid; when set to false, all editing is done via forms
   * Make checkcolumn respect the `read_only` option
+  * Column config option `editable` has been removed, use `read_only` instead
 
 * bug fix
   * Fix one-to-many association method falling back to 'id'